+++ src/Enums/KnowledgeScope.php
@@ -9,6 +9,7 @@ enum KnowledgeScope: string
     case GlobalShared = 'global_shared';
     case TenantPublic = 'tenant_public';
     case TenantPrivate = 'tenant_private';
+    case WorkspacePrivate = 'workspace_private';
     case UserPrivate = 'user_private';
     case SubscriptionLimited = 'subscription_limited';
 }